Fixed 32-bit compiling of MP4 output

+++ src/output/output_progressive_mp4.cpp
@@ -619,7 +619,7 @@ namespace Mist {
     //Fun fact! Firefox cares about the ordering here.
     //It doesn't care about the order or track IDs in the header.
     //But - the first TRAF must be a video TRAF, if video is present.
-    std::deque<std::map<long unsigned int, fragSet>::iterator> sortedTracks;
+    std::deque<std::map<size_t, fragSet>::iterator> sortedTracks;
     for (std::map<size_t, fragSet>::iterator it = currentPartSet.begin(); it != currentPartSet.end(); it++) {
       if (myMeta.tracks[it->first].type == "video"){
         sortedTracks.push_front(it);
@@ -631,9 +631,9 @@ namespace Mist {
       }
     }
 
-    for (std::deque<std::map<long unsigned int, fragSet>::iterator>::iterator ti = sortedTracks.begin(); ti != sortedTracks.end(); ++ti) {
-      std::map<long unsigned int, fragSet>::iterator & it = *ti;
-      unsigned int tid = it->first;
+    for (std::deque<std::map<size_t, fragSet>::iterator>::iterator ti = sortedTracks.begin(); ti != sortedTracks.end(); ++ti) {
+      std::map<size_t, fragSet>::iterator & it = *ti;
+      size_t tid = it->first;
       DTSC::Track & thisTrack = myMeta.tracks[tid];
       MP4::TRAF trafBox;
       MP4::TFHD tfhdBox;
